Refactors payment transaction check parameter handling and error reporting

+++ prime-modules/src/main/kotlin/org/ostelco/prime/paymentprocessor/core/Model.kt
@@ -32,4 +32,4 @@ data class InvoiceInfo(val id: String)
 
 data class InvoicePaymentInfo(val id: String, val chargeId: String)
 
-data class PaymentTransactionInfo(val id: String, val amount: Long, val currency: String, val created: Long, val refunded: Boolean, val details: Map<String, Any>)
+data class PaymentTransactionInfo(val id: String, val amount: Int, val currency: String, val created: Long, val refunded: Boolean, val details: Map<String, Any>)
